Delete files when corresponding db entry is removed
